+++ plugins/auth-backend/CHANGELOG.md
@@ -1,5 +1,17 @@
 # @backstage/plugin-auth-backend
 
+## 0.2.3
+
+### Patch Changes
+
+- Updated dependencies [1166fcc36]
+- Updated dependencies [bff3305aa]
+- Updated dependencies [1185919f3]
+- Updated dependencies [b47dce06f]
+  - @backstage/catalog-model@0.3.0
+  - @backstage/backend-common@0.3.1
+  - @backstage/catalog-client@0.3.1
+
 ## 0.2.2
 
 ### Patch Changes