+++ plugins/rollbar/CHANGELOG.md
@@ -1,5 +1,16 @@
 # @backstage/plugin-rollbar
 
+## 0.2.3
+
+### Patch Changes
+
+- Updated dependencies [475fc0aaa]
+- Updated dependencies [1166fcc36]
+- Updated dependencies [1185919f3]
+  - @backstage/core@0.3.2
+  - @backstage/catalog-model@0.3.0
+  - @backstage/plugin-catalog@0.2.3
+
 ## 0.2.2
 
 ### Patch Changes